+++ packages/@tinacms/app/appFiles/src/preview.tsx
@@ -18,7 +18,7 @@ import { useCMS, defineStaticConfig } from 'tinacms'
 type Config = Parameters<typeof defineStaticConfig>[0]
 
 type PostMessage = {
-  type: 'open' | 'close'
+  type: 'open' | 'close' | 'isEditMode'
   id: string
   data: object
 }
@@ -97,6 +97,13 @@ const QueryMachine = (props: {
 
   React.useEffect(() => {
     if (props.iframeRef.current) {
+      window.addEventListener('message', (event: MessageEvent<PostMessage>) => {
+        if (event?.data?.type === 'isEditMode') {
+          props.iframeRef?.current?.contentWindow?.postMessage({
+            type: 'tina:editMode',
+          })
+        }
+      })
       send({ type: 'IFRAME_MOUNTED', value: props.iframeRef.current })
       if (props.payload.type === 'open') {
         send({ type: 'ADD_QUERY', value: props.payload })