+++ packages/@tinacms/toolkit/src/packages/fields/plugins/MdxFieldPlugin/plate/plugins/create-mdx-plugins/component.tsx
@@ -215,8 +215,8 @@ const DotMenu = ({ onOpen, onRemove }) => {
         leaveFrom="transform opacity-100 scale-100"
         leaveTo="transform opacity-0 scale-95"
       >
-        <Popover.Panel>
-          <div className="z-30 origin-top-right absolute right-0 mt-2 -mr-1 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 focus:outline-none">
+        <Popover.Panel className="z-30 absolute origin-top-right right-0">
+          <div className="mt-2 -mr-1 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 focus:outline-none">
             <div className="py-1">
               <span
                 onClick={onOpen}