+++ examples/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,17 @@
 # e2e-next
 
+## 0.1.3
+
+### Patch Changes
+
+- tinacms@0.69.22
+
+## 0.1.2
+
+### Patch Changes
+
+- tinacms@0.69.21
+
 ## 0.1.1
 
 ### Patch Changes