+++ packages/tinacms/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tinacms
 
+## 0.69.22
+
+### Patch Changes
+
+- Updated dependencies [6ce6085e8]
+  - @tinacms/toolkit@0.58.4
+
+## 0.69.21
+
+### Patch Changes
+
+- Updated dependencies [48de42bc0]
+  - @tinacms/toolkit@0.58.3
+
 ## 0.69.20
 
 ### Patch Changes